:recycle: api: use service to check if session is scheduled in the past
