+++ api/lib/domain/services/sessions-import-validation-service.js
@@ -1,0 +1,13 @@
+const { UnprocessableEntityError } = require('../../application/http-errors');
+const Session = require('../models/Session');
+
+module.exports = {
+  async validate({ sessions }) {
+    sessions.forEach((session) => {
+      const domainSession = new Session(session);
+      if (domainSession.isSessionScheduledInThePast()) {
+        throw new UnprocessableEntityError(`Une session ne peut pas être programmée dans le passé`);
+      }
+    });
+  },
+};
